Move PositionalLayout#main() into :tools as PositionalLayoutDemo
